% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manning_baul.R
\name{manning_baul}
\alias{manning_baul}
\title{Cálculo hidráulico de un túnel de sección baúl cuadrada o no cuadrada}
\usage{
manning_baul(Q, K, n, So, x = NULL, seccion = c("cuadrada", "no_cuadrada"))
}
\arguments{
\item{Q}{numeric. Caudal en m3/s}

\item{K}{numeric. Relación tirante/radio}

\item{n}{numeric. Coeficiente de rugosidad de Manning}

\item{So}{numeric. Pendiente del túnel en m/m}

\item{x}{numeric. Relación altura/radio}

\item{seccion}{character. Tipo de sección (cuadrada o no_cuadrada)}
}
\value{
La función genera una lista con  los siguientes parámetros:
\itemize{
\item \code{radio}: radio de la sección semicircular del túnel tipo baúl.
\item \code{perim}: perímetro mojado.
\item \code{area}: área mojada.
\item \code{radio_h}: radio hidráulico.
\item \code{tirante}: tirante.
\item \code{velocidad}: velocidad del flujo.
\item \code{altura}: altura total del túnel.
\item \code{ancho}: ancho del túnel.
\item \code{BL}: altura de túnel no cubierta por agua.
\item \code{area_exc}: área de excavación del túnel.
}
}
\description{
Esta herramienta calcula la geometría y condiciones hidráulicas de un túnel de sección
baúl funcionando a sección parcial.
}
\section{Notes}{

Para \code{seccion="cuadrada"} no es necesario especificar \code{x}. Por el contrario,
este parámetro debe ser especificado si \verb{seccion="no_cuadrada}.
}

\examples{
# Túnel baúl cuadrado
Q <- 3
So <- 0.0004
n <- 0.04
K <- 1.5
manning_baul(Q = Q, K = K, n = n, So = So, seccion = "cuadrada")
# Túnel baúl no cuadrado
Q <- 3
So <- 0.0004
n <- 0.04
K <- 3.2
x <- 2.7
manning_baul(Q = Q, K = K, n = n, So = So, x = x, seccion = "no_cuadrada")

}
